ECM2434_A_2_202425/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
packaging==24.2
PyJWT==2.9.0
sqlparse==0.5.3
gunicorn==20.1.0
django
gunicorn
djangorestframework
pillow  # If you're handling images
django-cors-headers  # If you're using CORSwhitenoise>=6.0.0
whitenoise==6.5.0